+++ include/HashSet.h
@@ -85,7 +85,7 @@ namespace Lucene
 		
 		virtual int32_t hashCode()
 		{
-			return (int32_t)setContainer.get();
+			return (int32_t)(int64_t)setContainer.get();
 		}
 		
 		operator bool() const
