+++ util/FieldCacheSanityChecker.cpp
@@ -54,7 +54,7 @@ namespace Lucene
                 continue;
             
             ReaderFieldPtr rf(newLucene<ReaderField>(item->getReaderKey(), item->getFieldName()));
-            int32_t valId = (int32_t)val.get();
+            int32_t valId = (int32_t)(int64_t)val.get();
             
             // indirect mapping, so the MapOfSet will dedup identical valIds for us
             valIdToItems.put(valId, item);
